+++ adapters/celestia/src/verifier/mod.rs
@@ -45,14 +45,14 @@ impl BlobReaderTrait for BlobWithSender {
         self.blob.accumulator()
     }
 
+    fn total_len(&self) -> usize {
+        self.blob.total_len()
+    }
+
     #[cfg(feature = "native")]
     fn advance(&mut self, num_bytes: usize) -> &[u8] {
         self.blob.advance(num_bytes);
         self.verified_data()
-    }
-
-    fn total_len(&self) -> usize {
-        self.blob.total_len()
     }
 }
